@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { getVersions } from '../../actions/versions';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import paths from '../../util/paths';
+import {path} from '../../util/articles';
 
 /**
  * Version Navigation Component.
@@ -25,7 +25,7 @@ class Versions extends Component {
 
     onChangeVersion(e) {
         const version =  e.target.value === 'latest' ?  '' : e.target.value;
-        window.location.href = paths.concat(this.state.siteroot, version);
+        window.location.href = path.concat(this.state.siteroot, version);
     }
 
     componentWillReceiveProps(props) {
@@ -33,7 +33,7 @@ class Versions extends Component {
     }
 
     componentWillMount(){
-        this.props.getVersions(paths.concat(this.state.siteroot, 'versions.json'));
+        this.props.getVersions(path.concat(this.state.siteroot, 'versions.json'));
     }
 
     render() {
